# packages/contracts/openapi-cims-production-v0.18.yaml
openapi: 3.1.0
info:
  title: CIMS Production Manual Hearing Intake API
  version: 0.18.0
  description: Compliance-first orchestration API with manual initial hearing intake by Panitera Pengganti, maker-checker
    activation, PostgreSQL persistence, and a disabled-by-default foundation for future read-only database import.
servers:
- url: /api/v1
security:
- bearerAuth: []
...
paths:
...
  /electronic-hearing-requests:
    post:
      tags:
      - determinations
      operationId: createElectronicHearingRequest
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - hearing_id
              - requested_mode
              - reason
              properties:
                hearing_id:
                  type: string
                requested_mode:
                  type: string
                  enum:
                  - ELECTRONIC
                  - HYBRID
                reason:
                  type: string
                  minLength: 10
      responses:
        '201':
          description: Request created
        '403':
          $ref: '#/components/responses/Forbidden'
  /judicial-determinations:
    post:
      tags:
      - determinations
      operationId: createJudicialDetermination
      parameters:
      - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - hearing_id
              - decision
              - official_reference
              - reason
              properties:
                hearing_id:
                  type: string
                decision:
                  type: string
                  enum:
                  - APPROVED
                  - REJECTED
                official_reference:
                  type: string
                reason:
                  type: string
      responses:
        '201':
          description: Versioned judicial determination created
...
  /hearings/{hearingId}/participants:
    get:
      tags:
      - participants
      operationId: listParticipants
      parameters:
      - $ref: '#/components/parameters/HearingId'
      responses:
        '200':
          description: Participant list with protected identity masking
    post:
      tags:
      - participants
      operationId: registerParticipant
      parameters:
      - $ref: '#/components/parameters/HearingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterParticipant'
      responses:
        '201':
          description: Participant registered with encrypted sensitive fields
  /hearings/{hearingId}/participants/{participantId}/join-token:
    post:
      tags:
      - participants
      operationId: issueJoinToken
      parameters:
      - $ref: '#/components/parameters/HearingId'
      - $ref: '#/components/parameters/ParticipantId'
      responses:
        '201':
          description: Single-use plaintext token returned once
  /public/join-tokens/exchange:
    post:
      tags:
      - participants
      security: []
      operationId: exchangeJoinToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - token
              properties:
                token:
                  type: string
                  minLength: 32
      responses:
        '201':
          description: Waiting-room session created
        '410':
          description: Expired
          consumed: null
          or revoked token: null
  /hearings/{hearingId}/attendance:
    get:
      tags:
      - participants
      operationId: getAttendance
      parameters:
      - $ref: '#/components/parameters/HearingId'
      responses:
        '200':
          description: Append-only attendance timeline
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    HearingId:
      name: hearingId
      in: path
      required: true
      schema:
        type: string
    ParticipantId:
      name: participantId
      in: path
      required: true
      schema:
        type: string
...
    CorrelationId:
      name: x-correlation-id
      in: header
      required: false
      schema:
        type: string
...
  schemas:
...
    RegisterParticipant:
      type: object
      required:
      - role
      - displayName
      - protectedIdentity
      properties:
        role:
          type: string
          enum:
          - JUDGE
          - COURT_CLERK
          - PROSECUTOR
          - DEFENDANT
          - ADVOCATE
          - WITNESS
          - EXPERT
          - INTERPRETER
          - CORRECTIONS_OFFICER
          - IT_OPERATOR
        displayName:
          type: string
        alias:
          type: string
        organizationId:
          type: string
        contactEmail:
          type: string
          format: email
        protectedIdentity:
          type: boolean
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            correlation_id:
              type: string
            details: {}
...
  responses:
    Forbidden:
      description: RBAC, ABAC, or PostgreSQL RLS denied access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...